Cute and Spooky
Accepts user input "spirit animal" and combines that with other predetermined JSON data to generate a "cute and spooky" halloween themed image from a ML model. New images are added to the canvas alongside other previously generated images.

The program commits the input, the URL of the response as well as randomly generated coordinates to a LOWDB database. Upon every launch of the program, the program will iterate over the db and animate to the canvas each one of the user generated images.